splitting out train/test scripts and experiment bash script
